+++ src/test/java/org/scion/jpan/ProtobufPathDemo.java
@@ -23,7 +23,7 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import org.scion.jpan.demo.DemoConstants;
 import org.scion.jpan.internal.paths.ControlServiceGrpc;
-import org.scion.jpan.internal.paths.PathBuilder;
+import org.scion.jpan.internal.paths.Segments;
 import org.scion.jpan.proto.daemon.Daemon;
 import org.scion.jpan.testutil.TestUtil;
 
@@ -149,7 +149,7 @@ public class ProtobufPathDemo {
     ScionService csService =
         Scion.newServiceWithTopologyFile("topologies/tiny4/ASff00_0_112/topology.json");
     ControlServiceGrpc cs = PackageVisibilityHelper.getControlService(csService);
-    List<PathMetadata> paths = PathBuilder.getPathsCS(cs, csService.getLocalAS(), srcIA, dstIA);
+    List<PathMetadata> paths = Segments.getPathsCS(cs, csService.getLocalAS(), srcIA, dstIA);
     System.out.println("Paths found: " + paths.size());
     for (PathMetadata path : paths) {
       System.out.println("Path:  exp=" + path.getExpiration() + "  mtu=" + path.getMtu());